Revert "74x: Additional features needed for P5 visualization"
