+++ Fireworks/Core/doc/Core.doc
@@ -90,6 +90,7 @@
 - FWHFView
 - FWHLTTriggerTableView
 - FWHLTValidator
+- FWHandle
 - FWISpyView
 - FWIntValueListener
 - FWIntValueListenerBase
