+++ Fireworks/Core/interface/FW3DView.h
@@ -1,4 +1,4 @@
-#ifndef Fireworks_Core_FW3DView_h
+#ifndef Fireworks_Core_FW3DEnergyView_h
 #define Fireworks_Core_FW3DView_h
 // -*- C++ -*-
 //
