+++ Fireworks/Core/interface/FWGUIEventFilter.h
@@ -1,7 +1,3 @@
-
-#ifndef Fireworks_Core_GUIEventFilter_h
-#define Fireworks_Core_GUIEventFilter_h
-
 #include <list>
 
 #include "TGFrame.h"
@@ -83,4 +79,3 @@ private:
    CmsShowNavigator*    m_navigator;
 };
 
-#endif
