+++ Fireworks/Core/scripts/cmsShow
@@ -8,6 +8,8 @@ checkOS()
 
    osname=`uname -s`
    if [ f$osname = f"Darwin" ]; then
+
+       gdbCmd="lldb --"
        arch=`sysctl hw | grep 64bit | awk '{print $2;}'`
        if [ $arch -eq 0 ]; then
            echo "###################################"
@@ -111,15 +113,6 @@ setupEnv()
    fi
 }
 
-printEnv()
-{
-   # Needed only for debugging of a standalone version
-   for i in CMSSW_BASE CMSSW_SEARCH_PATH CMSSW_DATA_PATH LD_LIBRARY_PATH  DYLD_FALLBACK_LIBRARY_PATH PATH SHELLDIR; do
-       echo export $i=`eval echo \\$$i`
-   done
-   exit;
-}
-
 parseOptions()
 {
    # We store in originalOptions all the options that need to be passed to the
@@ -189,11 +182,8 @@ fi
 
 # Developers cases -- run from gdb or valgrind.
 if [ "X$debuginteractiveflag" = Xon ] ; then
-  if [[ `uname` == 'Linux' ]]; then
-      gdbCmd="gdb --args $SHELLDIR/cmsShow.exe $originalOptions"
-  else
-      gdbCmd="lldb $SHELLDIR/cmsShow.exe $originalOptions"
-  fi
+  gdbCmd="gdb --args " "$SHELLDIR/cmsShow.exe" $originalOptions
+  if-shell 'test "$(uname)" = "Darwin"' gdbCmd="lldb -- "
   exec $gdbCmd 
 elif [ "X$valgrindflag" = Xon ] ; then
   # MT: This should supposedly get rid of ROOT dicitonary reports etc. but it
