+++ Fireworks/Core/src/CmsShowMain.h
@@ -97,7 +97,6 @@ private:
    virtual void autoLoadNewEvent();
    virtual void checkPosition();
    virtual void stopPlaying();
-    virtual void checkKeyBindingsOnPLayEventsStateChanged();
 
    void reachedEnd();
    void reachedBeginning();
