+++ Fireworks/Core/test/unittest_fwconfiguration.cc
@@ -94,7 +94,7 @@ BOOST_AUTO_TEST_CASE( fwconfiguration )
    BOOST_CHECK(found->value()==kValue);
    BOOST_CHECK_THROW(config.valueForKey("blah"), std::runtime_error);
    
-   FWConfiguration::streamTo(std::cout, topConfig, "top");
+   streamTo(std::cout, topConfig, "top");
 
    //Test manager
    std::auto_ptr<Conf> pConf(new Conf() );
