+++ Fireworks/Muons/plugins/FWMuonRhoPhiProxyBuilder.cc
@@ -6,8 +6,6 @@
 //
 
 #include "Fireworks/Core/interface/FWSimpleProxyBuilderTemplate.h"
-#include "Fireworks/Core/interface/FWProxyBuilderConfiguration.h"
-#include "Fireworks/Core/interface/FWEventItem.h"
 #include "Fireworks/Muons/interface/FWMuonBuilder.h"
 #include "DataFormats/MuonReco/interface/Muon.h"
 
@@ -16,8 +14,6 @@ class FWMuonRhoPhiProxyBuilder : public FWSimpleProxyBuilderTemplate<reco::Muon>
 public:
    FWMuonRhoPhiProxyBuilder( void ) {}
    virtual ~FWMuonRhoPhiProxyBuilder( void ) {}
-
-   virtual void setItem(const FWEventItem* iItem);
 
    REGISTER_PROXYBUILDER_METHODS();
 
@@ -37,22 +33,9 @@ private:
 };
 
 void
-FWMuonRhoPhiProxyBuilder::setItem(const FWEventItem* iItem)
-{
-   FWProxyBuilderBase::setItem(iItem);
-   
-   if (iItem) {
-      iItem->getConfig()->assertParam("LineWidth", long(1), long(1), long(4));
-   }
-}
-
-void
 FWMuonRhoPhiProxyBuilder::build( const reco::Muon& iData, unsigned int iIndex,
                                  TEveElement& oItemHolder, const FWViewContext* ) 
 {
-   int width = item()->getConfig()->value<long>("LineWidth");
-   m_builder.setLineWidth(width);
-
    // To build in RhoPhi we should simply disable the Endcap drawing
    // by passing a false flag to a muon builder:
    m_builder.buildMuon( this, &iData, &oItemHolder, false, false );
